+++ moview/modules/initial_input/initial_question_giver.py
@@ -8,7 +8,7 @@ from langchain.prompts.chat import (
     HumanMessagePromptTemplate
 )
 from moview.modules.prompt_loader.prompt_loader import PromptLoader
-from moview.envrionment.llm_factory import LLMModelFactory
+from moview.environment.llm_factory import LLMModelFactory
 
 
 class InitialQuestionParseError(Exception):
